+++ js/lib/image_canvas.js
