+++ js/lib/index.js
@@ -8,5 +8,5 @@
 __webpack_public_path__ = document.querySelector('body').getAttribute('data-base-url') + 'nbextensions/yt-jscanvas/';
 
 // Export widget models and views, and the npm package version number.
-module.exports = require('./example.js');
+module.exports = require('./image_canvas.js');
 module.exports['version'] = require('../package.json').version;
