+++ yt_pycanvas/__init__.py
@@ -1,6 +1,6 @@
 from ._version import version_info, __version__
 
-from .example import *
+from .image_canvas import *
 
 def _jupyter_nbextension_paths():
     return [{
